+++ src/main/java/firok/tiths/common/Alloys.java
@@ -11,7 +11,7 @@ public class Alloys
 {
 	public static void registerAlloys()
 	{
-		TinkersThings.log("register alloys...");
+//		TinkersThings.log("register alloys...");
 		// 奢华合金
 		registerAlloy(
 				new FluidStack(Fluids.moltenRoyalAlloy, VALUE_Ingot*3),