+++ src/main/java/firok/tiths/traits/TraitBirefringent.java
@@ -12,7 +12,7 @@ import static firok.tiths.util.Predicates.canTrigger;
 // 双折射 (双折)
 public class TraitBirefringent extends AbstractTrait
 {
-	public static final float rate=0.3f;
+	public static final float rate=0.3f; // 双折几率
 	public TraitBirefringent()
 	{
 		super(nameTraitBirefringent, colorTraitBirefringent);
@@ -22,7 +22,7 @@ public class TraitBirefringent extends AbstractTrait
 	public void afterHit(ItemStack tool, EntityLivingBase player, EntityLivingBase target, float damageDealt, boolean wasCritical, boolean wasHit)
 	{
 		super.afterHit(tool, player, target, damageDealt, wasCritical, wasHit);
-		if(wasHit && player.isServerWorld() && target.isEntityAlive() && canTrigger(player.world,0.3f))
+		if(wasHit && player.isServerWorld() && target.isEntityAlive() && canTrigger(player.world,rate))
 		{
 			target.hurtResistantTime=0;
 			target.attackEntityFrom(BirefringentDamage,damageDealt);